This adds the AES flag for Clang, allowing to build fwZip successfully. 

Initially, the MR was opened to fix this issue but actually many targets failed to compile with Clang, thus other minor fixes are included.

Closes #482 
Merge branch '482-fwzip-fail-to-compile-with-clang' into 'dev'
See merge request Sight/sight!401

+++ SrcLib/core/fwDataIO/src/fwDataIO/reader/GzBufferImageReader.cpp
@@ -86,7 +86,7 @@ void GzBufferImageReader::read()
     while ( (uncompressedbytesreaded =
                  gzread(rawFile, ptr + readBytes, static_cast<unsigned int>(imageSizeInBytes - readBytes))) > 0 )
     {
-        readBytes += uncompressedbytesreaded;
+        readBytes += static_cast<size_t>(uncompressedbytesreaded);
     }
 
     gzclose(rawFile);
